Add files via upload
This is a program that first prompts the user to choose between checking a single number or a series of numbers. If the user selects "1", the program prompts for a single number and checks whether it is prime or not. If the user selects "2", the program prompts for a series of numbers and finds the prime numbers from the input. The results are then printed to the console. If the user enters an invalid choice, an error message is displayed.